ensure this script is a no-op when the ENV vars aren't set
Sometimes, configuring ssh via ENV vars is a helpful option.

Other times, users prefer to carefully construct the contents of ~/.ssh
via other means that this script can interfere with. For example,
non-root pods on kubernetes often don't have $HOME set, but ssh config
files might be mounted into the filesystem manually.

Without $HOME, t he mkdir and chmod at the top of this script try to
operate on //.ssh, which usually raises a permissions error.

This adds a short circuit to the script so none of the logic executes if
all the environment variables it looks for are unset. The giant new
conditional is ugly, but maybe it's worthwhile?